backend/requirements.txt
fastapi
uvicorn
pydantic
'pydantic[email]'
python-multipart
httpx
python-dotenv














